+++ lib/redmine_lightbox/patches/wiki_helper_patch.rb
@@ -1,0 +1,6 @@
+require_dependency 'application_helper'
+require_dependency 'wiki_helper'
+
+unless WikiHelper.included_modules.include?(ApplicationHelper)
+  WikiHelper.send(:include, ApplicationHelper)
+end
